# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ContentCloud Private Transport
  version: 1.0.0
  description: >-
    Private same-origin Web BFF, public review BFF, and CLI transport. This is not
    a public SDK. Agents, renderers, scripts, and CI use the contentcloud CLI.
    The cloud control plane never accepts LLM prompts, model credentials, or
    arbitrary execution requests.
servers:
  - url: /api
...
paths:
...
  /bff/projects/{project_id}/sources/search:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      tags: [Web BFF]
      operationId: searchSources
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [query]
              properties:
                query: {type: string, minLength: 2, maxLength: 500}
                limit: {type: integer, minimum: 1, maximum: 20, default: 10}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/projects/{project_id}/sources/fetch:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      tags: [Web BFF]
      operationId: fetchSource
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [url]
              properties:
                url: {type: string, format: uri, pattern: '^https?://'}
                name: {type: string}
                source_type: {type: string, default: web_fetch}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/projects/{project_id}/sources/upload:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      tags: [Web BFF]
      operationId: uploadSource
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, name, source_type]
              properties:
                file: {type: string, format: binary}
                name: {type: string}
                source_type: {type: string}
                file_type: {type: string, description: Optional MIME override when the client cannot provide the file MIME.}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
...
  /bff/sources/{source_id}/revisions/upload:
    parameters: [{name: source_id, in: path, required: true, schema: {type: string, format: uuid}}]
    post:
      tags: [Web BFF]
      operationId: uploadSourceRevision
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
                file_type: {type: string, description: Optional MIME override when the client cannot provide the file MIME.}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
...
  /bff/projects/{project_id}/knowledge-objects:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get: *jsonRead
    post:
      tags: [Web BFF]
      operationId: createKnowledgeObject
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateKnowledgeObjectInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/knowledge-objects/{id}/transitions:
    parameters: [{name: id, in: path, required: true, schema: {type: string, minLength: 1}}]
    post:
      tags: [Web BFF]
      operationId: transitionKnowledgeObject
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/ReviewKnowledgeObjectInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /bff/projects/{project_id}/knowledge-packs:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get: *jsonRead
    post:
      tags: [Web BFF]
      operationId: createKnowledgePack
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateKnowledgePackInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: cc_session}
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    ProjectID: {name: project_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  schemas:
...
    CreateKnowledgeObjectInput:
      type: object
      additionalProperties: false
      required: [object_type, layer]
      properties:
        id: {type: string, minLength: 1}
        object_type: {$ref: '#/components/schemas/KnowledgeObject/properties/object_type'}
        layer: {$ref: '#/components/schemas/KnowledgeObject/properties/layer'}
        status: {$ref: '#/components/schemas/KnowledgeObject/properties/status'}
        title: {type: string}
        statement: {type: string}
        payload: {type: object, additionalProperties: true}
        dimensions: {type: array, uniqueItems: true, items: {type: string}}
        allowed_channels: {type: array, uniqueItems: true, items: {type: string}}
        evidence_refs: {type: array, uniqueItems: true, items: {type: string}}
        relation_refs: {type: array, uniqueItems: true, items: {type: string}}
        rights_refs: {type: array, uniqueItems: true, items: {type: string}}
        conflict_refs: {type: array, uniqueItems: true, items: {type: string}}
        decision_ref: {type: string}
        next_action: {type: string}
        impact: {type: string}
        valid_from: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
    ReviewKnowledgeObjectInput:
      type: object
      additionalProperties: false
      required: [expected_version, expected_digest, decision, reason]
      properties:
        expected_version: {type: integer, minimum: 1}
        expected_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        decision: {type: string, enum: [approve, reject]}
        reason: {type: string, minLength: 1}
...
    KnowledgePackObjectRef:
      type: object
      additionalProperties: false
      required: [object_id, version]
      properties:
        object_id: {type: string, minLength: 1}
        version: {type: integer, minimum: 0, description: Zero resolves to the latest version when publishing.}
    KnowledgeQueryPolicy:
      type: object
      additionalProperties: false
      properties:
        eligible_statuses: {type: array, uniqueItems: true, items: {type: string, enum: [verified, approved, valid, active]}}
        allowed_object_types: {type: array, uniqueItems: true, items: {type: string}}
        require_evidence: {type: boolean, description: Enforced true by the service.}
        block_on_conflict: {type: boolean, description: Enforced true by the service.}
        block_on_rights_failure: {type: boolean, description: Enforced true by the service.}
    CreateKnowledgePackInput:
      type: object
      additionalProperties: false
      required: [name, purpose, object_refs]
      properties:
        id: {type: string, minLength: 1}
        name: {type: string, minLength: 1}
        purpose: {type: string, minLength: 1}
        object_refs: {type: array, minItems: 1, uniqueItems: true, items: {$ref: '#/components/schemas/KnowledgePackObjectRef'}}
        query_policy: {$ref: '#/components/schemas/KnowledgeQueryPolicy'}
...
    SuccessEnvelope:
      type: object
      required: [ok, command, request_id, data, meta]
      properties:
        ok: {const: true}
        command: {type: string}
        request_id: {type: string}
        data: {}
        meta: {type: object}
    ErrorEnvelope:
      type: object
      required: [ok, command, request_id, error]
      properties:
        ok: {const: false}
        command: {type: string}
        request_id: {type: string}
        error:
          type: object
          required: [type, subtype, code, message, retryable]
          properties:
            type: {type: string}
            subtype: {type: string}
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            hint: {type: string}
            details: {}
  responses:
    Success:
      description: Stable success envelope
      content:
        application/json: {schema: {$ref: '#/components/schemas/SuccessEnvelope'}}
    Error:
      description: Stable structured error envelope
      content:
        application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}